distsql: move generic and no-op processor code to new file
