+++ pkg/sql/distsqlrun/base.go
@@ -17,13 +17,11 @@
 package distsqlrun
 
 import (
-	"sync"
 	"sync/atomic"
 
 	"golang.org/x/net/context"
 
 	"github.com/cockroachdb/cockroach/pkg/sql/sqlbase"
-	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/cockroach/pkg/util/tracing"
 	basictracer "github.com/opentracing/basictracer-go"
 	opentracing "github.com/opentracing/opentracing-go"
@@ -66,49 +64,6 @@ type RowSource interface {
 	// rows. Depending on the implementation, it may block.
 	// The caller must not modify the received row.
 	NextRow() (sqlbase.EncDatumRow, error)
-}
-
-// processor is a common interface implemented by all processors, used by the
-// higher-level flow orchestration code.
-type processor interface {
-	// Run is the main loop of the processor.
-	// If wg is non-nil, wg.Done is called before exiting.
-	Run(wg *sync.WaitGroup)
-}
-
-// noopProcessor is a processor that simply passes rows through from the
-// synchronizer to the router. It can be useful in the last stage of a
-// computation, where we may only need the synchronizer to join streams.
-type noopProcessor struct {
-	flowCtx *FlowCtx
-	input   RowSource
-	output  RowReceiver
-}
-
-var _ processor = &noopProcessor{}
-
-func newNoopProcessor(flowCtx *FlowCtx, input RowSource, output RowReceiver) *noopProcessor {
-	return &noopProcessor{flowCtx: flowCtx, input: input, output: output}
-}
-
-// Run is part of the processor interface.
-func (n *noopProcessor) Run(wg *sync.WaitGroup) {
-	if wg != nil {
-		defer wg.Done()
-	}
-	for {
-		row, err := n.input.NextRow()
-		if err != nil || row == nil {
-			n.output.Close(err)
-			return
-		}
-		if log.V(3) {
-			log.Infof(n.flowCtx.Context, "noop: pushing row %s", row)
-		}
-		if !n.output.PushRow(row) {
-			return
-		}
-	}
 }
 
 // StreamMsg is the message used in the channels that implement
